Move computeConflictedDirectoryPaths to ui/shared
Fixes Copilot review comment: the function was in ui/treemap but used by
both ui/treemap and ui/explorer, creating an unwanted cross-feature
dependency. Moving it to ui/shared/OverlayComputer.kt where the other
overlay computations live.
